+++ CHANGELOG.md
@@ -23,6 +23,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
 <!-- Types of changes: Added, Changed, Deprecated, Removed, Fixed -->
 
+## [0.9.0] - 18/09/2018
+**NOTE:** The chess.com analysis board does not support subvariations and hence will not be able to read any PGN4 with subvariations.
+### Added:
+- Option to use chess.com compatible FEN4 and PGN4
+- Player rating (editable)
+
+
 ## [0.8.0] - 03/09/2018
 ### Added:
 - Legal move indicators
@@ -108,6 +115,7 @@ Initial development release.
 
 
 <!-- Links to releases -->
+[0.9.0]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.9.0
 [0.8.0]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.8.0
 [0.7.0]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.7.0
 [0.6.1]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.6.1